# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HireTrack Job Portal API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  message: string;
  errors?: string[];
}

export interface MessageResponse {
  message: string;
}

export interface TokenResponse {
  accessToken: string;
  refreshToken: string;
}

export type RegisterBodyRole =
  (typeof RegisterBodyRole)[keyof typeof RegisterBodyRole];

export const RegisterBodyRole = {
  candidate: "candidate",
  recruiter: "recruiter",
} as const;

export interface RegisterBody {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  role: RegisterBodyRole;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface RefreshBody {
  refreshToken: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  candidate: "candidate",
  recruiter: "recruiter",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  accessToken: string;
  refreshToken: string;
}

export interface ExperienceItem {
  title: string;
  company: string;
  startDate: string;
  endDate?: string | null;
  current: boolean;
  description?: string | null;
}

export interface EducationItem {
  degree: string;
  institution: string;
  field: string;
  startYear: number;
  endYear?: number | null;
}

export interface Profile {
  id: number;
  userId: number;
  headline?: string | null;
  bio?: string | null;
  location?: string | null;
  phone?: string | null;
  skills: string[];
  experience: ExperienceItem[];
  education: EducationItem[];
  resumeUrl?: string | null;
  avatarUrl?: string | null;
  user: User;
}

export interface UpdateProfileBody {
  headline?: string;
  bio?: string;
  location?: string;
  phone?: string;
  skills?: string[];
  experience?: ExperienceItem[];
  education?: EducationItem[];
}

export interface UploadResumeBody {
  fileName: string;
  /** Base64 encoded PDF */
  fileData: string;
}

export interface ResumeUploadResponse {
  resumeUrl: string;
}

export interface Company {
  id: number;
  recruiterId: number;
  name: string;
  description?: string | null;
  website?: string | null;
  location?: string | null;
  industry?: string | null;
  size?: string | null;
  logoUrl?: string | null;
}

export interface UpdateCompanyBody {
  name?: string;
  description?: string;
  website?: string;
  location?: string;
  industry?: string;
  size?: string;
}

export type JobJobType = (typeof JobJobType)[keyof typeof JobJobType];

export const JobJobType = {
  "full-time": "full-time",
  "part-time": "part-time",
  contract: "contract",
  internship: "internship",
  remote: "remote",
} as const;

export interface Job {
  id: number;
  title: string;
  description: string;
  location: string;
  jobType: JobJobType;
  salaryMin?: number | null;
  salaryMax?: number | null;
  skills: string[];
  isActive: boolean;
  applicationCount: number;
  createdAt: string;
  company: Company;
  recruiterId: number;
}

export type CreateJobBodyJobType =
  (typeof CreateJobBodyJobType)[keyof typeof CreateJobBodyJobType];

export const CreateJobBodyJobType = {
  "full-time": "full-time",
  "part-time": "part-time",
  contract: "contract",
  internship: "internship",
  remote: "remote",
} as const;

export interface CreateJobBody {
  title: string;
  description: string;
  location: string;
  jobType: CreateJobBodyJobType;
  salaryMin?: number;
  salaryMax?: number;
  skills: string[];
}

export interface JobsListResponse {
  jobs: Job[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export type ApplicationStatus =
  (typeof ApplicationStatus)[keyof typeof ApplicationStatus];

export const ApplicationStatus = {
  applied: "applied",
  shortlisted: "shortlisted",
  rejected: "rejected",
  hired: "hired",
} as const;

export interface Application {
  id: number;
  candidateId: number;
  jobId: number;
  status: ApplicationStatus;
  coverLetter?: string | null;
  resumeUrl?: string | null;
  createdAt: string;
  updatedAt: string;
  job: Job;
  candidate: Profile;
}

export interface ApplyBody {
  jobId: number;
  coverLetter?: string;
  resumeUrl?: string;
}

export type UpdateStatusBodyStatus =
  (typeof UpdateStatusBodyStatus)[keyof typeof UpdateStatusBodyStatus];

export const UpdateStatusBodyStatus = {
  applied: "applied",
  shortlisted: "shortlisted",
  rejected: "rejected",
  hired: "hired",
} as const;

export interface UpdateStatusBody {
  status: UpdateStatusBodyStatus;
}

export type RecruiterDashboardApplicationsByStatus = {
  applied: number;
  shortlisted: number;
  rejected: number;
  hired: number;
};

export interface RecruiterDashboard {
  totalJobs: number;
  activeJobs: number;
  totalApplications: number;
  applicationsByStatus: RecruiterDashboardApplicationsByStatus;
  recentApplications: Application[];
  topJobs: Job[];
}

export interface CandidateDashboard {
  totalApplications: number;
  applied: number;
  shortlisted: number;
  rejected: number;
  hired: number;
  recentApplications: Application[];
}

export type ListJobsParams = {
  page?: number;
  limit?: number;
  keyword?: string;
  location?: string;
  skills?: string;
  jobType?: ListJobsJobType;
};

export type ListJobsJobType =
  (typeof ListJobsJobType)[keyof typeof ListJobsJobType];

export const ListJobsJobType = {
  "full-time": "full-time",
  "part-time": "part-time",
  contract: "contract",
  internship: "internship",
  remote: "remote",
} as const;